experiments/masterdp-chatbot/requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
flask==3.1.0
